Text related logic moved to separate behavior
